ALL-3028 Add Solana Rpc support
