+++ src/service/rpc/index.ts
@@ -1,6 +1,6 @@
 export * from './evm'
 export * from './evm/AbstractTronRpc'
 export * from './generic'
+export * from './other/AbstractSolanaRpc'
 export * from './other/AbstractXrpRpc'
-export * from './other/SolanaRpc'
 export * from './utxo'
